[Code] Refactor declarations into _ndforward.hpp
A drawback of the current code hierarchy is that there is no single source to check when looking up docstrings, or to quickly glance at available functions.

This refactor moves all declarations into _ndforward.hpp. This file is now the go-to reference to browse what the library has to offer.